requirements.txt
fastapi==0.54.1
uvicorn==0.11.5
python-multipart==0.0.5
pyjwt==1.7.1
passlib[bcrypt]==1.7.2
psycopg2-binary==2.8.5
pandas==1.0.3
gunicorn==20.0.4
gevent==20.4.0
boto3==1.13.1
botocore==1.16.1
aws-logging-handlers==2.0.3
testing.postgresql==1.3.0
